requirements.txt
youtube_dl==2020.1.24
APScheduler==3.7.0
git+https://github.com/Rapptz/discord-ext-menus
PyDrive==1.3.1
aiohttp==3.6.2
treelib==1.6.1
requests==2.22.0
aiohttp==3.6.2
discord==1.0.1
GitPython==3.1.14
python_dateutil==2.8.1
python-dateutil~=2.8.1
psutil~=5.8.0
bs4~=0.0.1
beautifulsoup4~=4.9.3